Backend/requirements.txt
tensorflow
numpy
Pillow
flask
keras
scikit-learn
flask-cors